moved topcontainer into profile component folder
